Rating and reviews working

+++ app/src/main/java/com/example/grindlyapp1/network/SubmitReviewRequest.kt
@@ -3,6 +3,6 @@ package com.example.grindlyapp1.network
 
 data class SubmitReviewRequest(
     val serviceId: String,
-    val rating: Int,
+    val rating: Double,
     val comment: String? = null
 )
